+++ lib/sbeu_reactive_pattern.dart
@@ -1,4 +1,2 @@
 export 'reactive/reactive_widget.dart';
 export 'reactive/reactive_wm.dart';
-export 'reactive/reactive_container.dart';
-export 'reactive/reactive_model.dart';